requirements.txt
# Power Demand Forecasting - Requirements
# Python 3.10+ | M1 MacBook Pro (MPS)

# ============================================
# Core Data Processing
# ============================================
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# ============================================
# Machine Learning
# ============================================
scikit-learn>=1.3.0

# ============================================
# Deep Learning (PyTorch with MPS support)
# ============================================
torch>=2.0.0
torchvision>=0.15.0

# ============================================
# Visualization
# ============================================
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# ============================================
# Jupyter/Colab Support
# ============================================
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.0.0

# ============================================
# Web Scraping (Weather Data Crawler)
# ============================================
selenium>=4.15.0
webdriver-manager>=4.0.0
requests>=2.31.0

# ============================================
# Utilities
# ============================================
python-dateutil>=2.8.0
tqdm>=4.65.0
PyYAML>=6.0.0
python-dotenv>=1.0.0

# ============================================
# Korean Holidays (for holiday features)
# ============================================
holidays>=0.25

# ============================================
# Logging & Monitoring
# ============================================
tensorboard>=2.13.0

# ============================================
# Testing
# ============================================
pytest>=7.4.0
pytest-cov>=4.1.0

# ============================================
# Code Quality (Optional)
# ============================================
# black>=23.0.0
# flake8>=6.0.0
# isort>=5.12.0

# ============================================
# Hyperparameter Tuning (Optional)
# ============================================
# optuna>=3.3.0
